Migrate buildSrc plugins to build-logic
