+++ build-logic/src/main/kotlin/velocity-checkstyle.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+    checkstyle
+}
+
+extensions.configure<CheckstyleExtension> {
+    configFile = rootProject.file("config/checkstyle/checkstyle.xml")
+    maxErrors = 0
+    maxWarnings = 0
+    toolVersion = libs.checkstyle.get().version.toString()
+}
