+++ build-logic/src/main/kotlin/velocity-spotless.gradle.kts
@@ -1,0 +1,15 @@
+import com.diffplug.gradle.spotless.SpotlessExtension
+import com.diffplug.gradle.spotless.SpotlessPlugin
+
+apply<SpotlessPlugin>()
+
+extensions.configure<SpotlessExtension> {
+    java {
+        if (project.name == "velocity-api") {
+            licenseHeaderFile(file("HEADER.txt"))
+        } else {
+            licenseHeaderFile(rootProject.file("HEADER.txt"))
+        }
+        removeUnusedImports()
+    }
+}
