+++ build.gradle.kts
@@ -1,28 +1,19 @@
-import com.velocitypowered.script.VelocityCheckstylePlugin
-import com.velocitypowered.script.VelocityPublishPlugin
-import com.velocitypowered.script.VelocitySpotlessPlugin
-
 plugins {
     `java-library`
+    id("velocity-checkstyle") apply false
+    id("velocity-spotless") apply false
 }
 
 subprojects {
     apply<JavaLibraryPlugin>()
 
-    apply<VelocityCheckstylePlugin>()
-    apply<VelocityPublishPlugin>()
-    apply<VelocitySpotlessPlugin>()
+    apply(plugin = "velocity-checkstyle")
+    apply(plugin = "velocity-spotless")
 
     java {
         toolchain {
             languageVersion.set(JavaLanguageVersion.of(11))
         }
-    }
-
-    repositories {
-        mavenCentral()
-        maven("https://s01.oss.sonatype.org/content/repositories/snapshots/") // adventure
-        maven("https://repo.papermc.io/repository/maven-public/")
     }
 
     dependencies {
