+++ native/build.gradle.kts
@@ -1,6 +1,6 @@
 plugins {
     `java-library`
-    `maven-publish`
+    id("velocity-publish")
 }
 
 dependencies {
