+++ settings.gradle.kts
@@ -1,3 +1,23 @@
+@file:Suppress("UnstableApiUsage")
+
+dependencyResolutionManagement {
+    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
+    repositories {
+        mavenCentral()
+        maven("https://s01.oss.sonatype.org/content/repositories/snapshots/") // adventure
+        maven("https://repo.papermc.io/repository/maven-public/")
+    }
+}
+
+
+pluginManagement {
+    includeBuild("build-logic")
+    repositories {
+        mavenCentral()
+        gradlePluginPortal()
+    }
+}
+
 plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.4.0"
 }
@@ -6,8 +26,8 @@ rootProject.name = "velocity"
 
 sequenceOf(
     "api",
-    "proxy",
     "native",
+    "proxy",
 ).forEach {
     val project = ":velocity-$it"
     include(project)
